requirements.txt
pillow
pyside6
natsort